separate packages to entity and dtos
