+++ src/main/java/com/votingapplication/repo/VotingRepo.java
@@ -1,6 +1,6 @@
 package com.votingapplication.repo;
 
-import com.votingapplication.resource.VoterEntity;
+import com.votingapplication.entity.VoterEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface VotingRepo extends JpaRepository<VoterEntity, Integer> {
